目前为止包含两个不完全的os_Debug的程序
程序基本已经完备，除了消息的反馈。包含两个不完全的os_Debug程序，将马上进行修正，然后最为example出现。所有的Serial.Print
和非必要的delay已经注释。接下来将进行1.返回值的增加；2.bluetooth连接的简化；3.两个os_Debug的完善

+++ Project_SoCMainFrame/Attached.cpp
@@ -9,6 +9,6 @@ int Devide()
 
 int Combine()
 {
-	whole=part1<<8+part2;
+	whole=(((int)part1)<<8)+(int)part2;
 	return 0;
 }
